TP4/requirements.txt
matplotlib>=3.3.4
numpy>=1.19.5
pandas>=1.1.5
PyYAML>=3.12
schema>=0.7.4
attrs>=20.3.0
scipy>=1.6.2
scikit-learn>=0.24.2